Improvements to API response output
